Move aggregatePlanNodeStats and deps to PlanNodeStatsSummarizer
