+++ presto-main/src/main/java/com/facebook/presto/sql/planner/planPrinter/PlanPrinter.java
@@ -16,16 +16,11 @@ package com.facebook.presto.sql.planner.planPrinter;
 import com.facebook.presto.Session;
 import com.facebook.presto.execution.StageInfo;
 import com.facebook.presto.execution.StageStats;
-import com.facebook.presto.execution.TaskInfo;
 import com.facebook.presto.metadata.Metadata;
 import com.facebook.presto.metadata.OperatorNotFoundException;
 import com.facebook.presto.metadata.Signature;
 import com.facebook.presto.metadata.TableHandle;
 import com.facebook.presto.metadata.TableLayout;
-import com.facebook.presto.operator.HashCollisionsInfo;
-import com.facebook.presto.operator.OperatorStats;
-import com.facebook.presto.operator.PipelineStats;
-import com.facebook.presto.operator.TaskStats;
 import com.facebook.presto.spi.ColumnHandle;
 import com.facebook.presto.spi.ConnectorTableLayoutHandle;
 import com.facebook.presto.spi.predicate.Domain;
@@ -100,12 +95,9 @@ import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Lists;
 import io.airlift.slice.Slice;
-import io.airlift.units.Duration;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Locale;
@@ -119,21 +111,15 @@ import static com.facebook.presto.execution.StageInfo.getAllStages;
 import static com.facebook.presto.spi.type.VarcharType.VARCHAR;
 import static com.facebook.presto.sql.planner.DomainUtils.simplifyDomain;
 import static com.facebook.presto.sql.planner.SystemPartitioningHandle.SINGLE_DISTRIBUTION;
+import static com.facebook.presto.sql.planner.planPrinter.PlanNodeStatsSummarizer.aggregatePlanNodeStats;
 import static com.facebook.presto.util.ImmutableCollectors.toImmutableList;
-import static com.facebook.presto.util.MoreMaps.mergeMaps;
 import static com.google.common.base.CaseFormat.UPPER_UNDERSCORE;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkState;
-import static com.google.common.collect.Iterables.getLast;
 import static com.google.common.collect.Iterables.getOnlyElement;
-import static com.google.common.collect.Lists.reverse;
-import static io.airlift.units.DataSize.Unit.BYTE;
-import static io.airlift.units.DataSize.succinctDataSize;
 import static java.lang.Double.isFinite;
 import static java.lang.String.format;
-import static java.util.Collections.emptyMap;
 import static java.util.Objects.requireNonNull;
-import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static java.util.stream.Collectors.toList;
 
 public class PlanPrinter
@@ -206,126 +192,6 @@ public class PlanPrinter
         }
 
         return builder.toString();
-    }
-
-    private static Map<PlanNodeId, PlanNodeStats> aggregatePlanNodeStats(StageInfo stageInfo)
-    {
-        Map<PlanNodeId, PlanNodeStats> aggregatedStats = new HashMap<>();
-        List<PlanNodeStats> planNodeStats = stageInfo.getTasks().stream()
-                .map(TaskInfo::getStats)
-                .flatMap(taskStats -> getPlanNodeStats(taskStats).stream())
-                .collect(toList());
-        for (PlanNodeStats stats : planNodeStats) {
-            aggregatedStats.merge(stats.getPlanNodeId(), stats, PlanNodeStats::merge);
-        }
-        return aggregatedStats;
-    }
-
-    private static List<PlanNodeStats> getPlanNodeStats(TaskStats taskStats)
-    {
-        // Best effort to reconstruct the plan nodes from operators.
-        // Because stats are collected separately from query execution,
-        // it's possible that some or all of them are missing or out of date.
-        // For example, a LIMIT clause can cause a query to finish before stats
-        // are collected from the leaf stages.
-        Map<PlanNodeId, Long> planNodeInputPositions = new HashMap<>();
-        Map<PlanNodeId, Long> planNodeInputBytes = new HashMap<>();
-        Map<PlanNodeId, Long> planNodeOutputPositions = new HashMap<>();
-        Map<PlanNodeId, Long> planNodeOutputBytes = new HashMap<>();
-        Map<PlanNodeId, Long> planNodeWallMillis = new HashMap<>();
-
-        Map<PlanNodeId, Map<String, OperatorInputStats>> operatorInputStats = new HashMap<>();
-        Map<PlanNodeId, Map<String, OperatorHashCollisionsStats>> operatorHashCollisionsStats = new HashMap<>();
-
-        for (PipelineStats pipelineStats : taskStats.getPipelines()) {
-            // Due to eventual consistently collected stats, these could be empty
-            if (pipelineStats.getOperatorSummaries().isEmpty()) {
-                continue;
-            }
-
-            Set<PlanNodeId> processedNodes = new HashSet<>();
-            PlanNodeId inputPlanNode = pipelineStats.getOperatorSummaries().iterator().next().getPlanNodeId();
-            PlanNodeId outputPlanNode = getLast(pipelineStats.getOperatorSummaries()).getPlanNodeId();
-
-            // Gather input statistics
-            for (OperatorStats operatorStats : pipelineStats.getOperatorSummaries()) {
-                PlanNodeId planNodeId = operatorStats.getPlanNodeId();
-
-                long wall = operatorStats.getAddInputWall().toMillis() + operatorStats.getGetOutputWall().toMillis() + operatorStats.getFinishWall().toMillis();
-                planNodeWallMillis.merge(planNodeId, wall, Long::sum);
-
-                // A pipeline like hash build before join might link to another "internal" pipelines which provide actual input for this plan node
-                if (operatorStats.getPlanNodeId().equals(inputPlanNode) && !pipelineStats.isInputPipeline()) {
-                    continue;
-                }
-                if (processedNodes.contains(planNodeId)) {
-                    continue;
-                }
-
-                operatorInputStats.merge(planNodeId,
-                        ImmutableMap.of(
-                                operatorStats.getOperatorType(),
-                                new OperatorInputStats(
-                                        operatorStats.getTotalDrivers(),
-                                        operatorStats.getInputPositions(),
-                                        operatorStats.getSumSquaredInputPositions())),
-                        PlanPrinter::mergeOperatorInputStatsMaps);
-
-                if (operatorStats.getInfo() instanceof HashCollisionsInfo) {
-                    HashCollisionsInfo hashCollisionsInfo = (HashCollisionsInfo) operatorStats.getInfo();
-                    operatorHashCollisionsStats.merge(planNodeId,
-                            ImmutableMap.of(
-                                    operatorStats.getOperatorType(),
-                                    new OperatorHashCollisionsStats(
-                                            hashCollisionsInfo.getWeightedHashCollisions(),
-                                            hashCollisionsInfo.getWeightedSumSquaredHashCollisions(),
-                                            hashCollisionsInfo.getWeightedExpectedHashCollisions())),
-                            PlanPrinter::mergeOperatorHashCollisionsStatsMaps);
-                }
-
-                planNodeInputPositions.merge(planNodeId, operatorStats.getInputPositions(), Long::sum);
-                planNodeInputBytes.merge(planNodeId, operatorStats.getInputDataSize().toBytes(), Long::sum);
-                processedNodes.add(planNodeId);
-            }
-
-            // Gather output statistics
-            processedNodes.clear();
-            for (OperatorStats operatorStats : reverse(pipelineStats.getOperatorSummaries())) {
-                PlanNodeId planNodeId = operatorStats.getPlanNodeId();
-
-                // An "internal" pipeline like a hash build, links to another pipeline which is the actual output for this plan node
-                if (operatorStats.getPlanNodeId().equals(outputPlanNode) && !pipelineStats.isOutputPipeline()) {
-                    continue;
-                }
-                if (processedNodes.contains(planNodeId)) {
-                    continue;
-                }
-
-                planNodeOutputPositions.merge(planNodeId, operatorStats.getOutputPositions(), Long::sum);
-                planNodeOutputBytes.merge(planNodeId, operatorStats.getOutputDataSize().toBytes(), Long::sum);
-                processedNodes.add(planNodeId);
-            }
-        }
-
-        List<PlanNodeStats> stats = new ArrayList<>();
-        for (Map.Entry<PlanNodeId, Long> entry : planNodeWallMillis.entrySet()) {
-            PlanNodeId planNodeId = entry.getKey();
-            stats.add(new PlanNodeStats(
-                    planNodeId,
-                    new Duration(planNodeWallMillis.get(planNodeId), MILLISECONDS),
-                    planNodeInputPositions.get(planNodeId),
-                    succinctDataSize(planNodeInputBytes.get(planNodeId), BYTE),
-                    // It's possible there will be no output stats because all the pipelines that we observed were non-output.
-                    // For example in a query like SELECT * FROM a JOIN b ON c = d LIMIT 1
-                    // It's possible to observe stats after the build starts, but before the probe does
-                    // and therefore only have wall time, but no output stats
-                    planNodeOutputPositions.getOrDefault(planNodeId, 0L),
-                    succinctDataSize(planNodeOutputBytes.getOrDefault(planNodeId, 0L), BYTE),
-                    operatorInputStats.get(planNodeId),
-                    // Only some operators emit hash collisions statistics
-                    operatorHashCollisionsStats.getOrDefault(planNodeId, emptyMap())));
-        }
-        return stats;
     }
 
     public static String textDistributedPlan(SubPlan plan, Metadata metadata, Session session)
@@ -1336,15 +1202,5 @@ public class PlanPrinter
         catch (Throwable throwable) {
             throw Throwables.propagate(throwable);
         }
-    }
-
-    private static <K> Map<K, OperatorInputStats> mergeOperatorInputStatsMaps(Map<K, OperatorInputStats> map1, Map<K, OperatorInputStats> map2)
-    {
-        return mergeMaps(map1, map2, OperatorInputStats::merge);
-    }
-
-    private static <K> Map<K, OperatorHashCollisionsStats> mergeOperatorHashCollisionsStatsMaps(Map<K, OperatorHashCollisionsStats> map1, Map<K, OperatorHashCollisionsStats> map2)
-    {
-        return mergeMaps(map1, map2, OperatorHashCollisionsStats::merge);
     }
 }
